+++ scripts/macros/classFeatures/druid/circleOfDreams/hearthOfMoonlightAndShadow.js
@@ -1,39 +1,20 @@
-import {chris} from '../../helperFunctions.js';
-
+import {chris} from '../../../../helperFunctions.js';
 async function item() {
     await warpgate.wait(50);
-    let effectUpdates = {
-        'flags': {
-            'dae': {
-                'transfer': false,
-                'showIcon': true,
-                'specialDuration': [
-                    'shortRest',
-                    'longRest'
-                ],
-                'stackable': 'noneName',
-                'macroRepeat': 'none'
-            }
-        }
-    }
-
-    let templateEffect = chrisPremades.helpers.findEffect(workflow.actor, workflow.item.name + ' Template');    
-    if (templateEffect) {
-        await chrisPremades.helpers.updateEffect(templateEffect, effectUpdates);
-    }
+    let effect = chrisPremades.helpers.findEffect(workflow.actor, workflow.item.name + ' Template');
+    if (!effect) return;
+    await effect.setFlag('dae', 'specialDuration', [
+        'shortRest',
+        'longRest'
+    ]);
 }
-
 async function left(template, token) {
-    const originActor = await fromUuid(template.flags["midi-qol"].actorUuid);
-    if (token.actor === originActor) {
-        const originItem = await fromUuid(template.flags["midi-qol"].itemUuid);
-        let effect = chris.findEffect(token.actor, originItem.name + ' Template');
-        if (effect) {
-            await chris.removeEffect(effect);
-        }
-    }
+    let originActor = await fromUuid(template.flags['midi-qol'].actorUuid);
+    if (token.actor != originActor) return;
+    let originItem = await fromUuid(template.flags['midi-qol'].itemUuid);
+    let effect = chris.findEffect(token.actor, originItem.name + ' Template');
+    if (effect) await chris.removeEffect(effect);
 }
-
 export let hearthOfMoonlightAndShadow = {
     'item': item,
     'left' : left,