+++ CHANGELOG.md
@@ -1,9 +1,13 @@
-# 0.10.14 Release:
-## Bug Fixes:
-- Manual Rolls
+# 0.10.15 Release:
+## New Content:
+- Circle of Dreams Druid: Balm of the Summer Court, Hearth of Moonlight and Shadow, and Hidden Paths (Thanks Sagatympana)
 ## Update Notes:
 - Many macros were adjusted for this update. There may be bugs, please report them on my Github page or Discord server.
 - This is a beta version for the 3.0.3 D&D system. Do not update to this version unless unless you have all the required up to date dependency modules.
+  
+# 0.10.14 Release:
+## Bug Fixes:
+- Manual Rolls
   
 # 0.10.13 Release:
 ## New Content: